requirements.txt
pandas==2.2.3
pyarrow==19.0.0
PyYAML==6.0.2
setuptools==75.1.0
wikitextparser==0.56.3
